[REF/#90] Response Request 모델 분리
